% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_wordcloud.R
\name{create_wordcloud}
\alias{create_wordcloud}
\title{create_wordcloud}
\usage{
create_wordcloud(lyrics, ngram = 1)
}
\arguments{
\item{lyrics}{JSON file with lyrics of DF_lyrics}

\item{ngram}{Show n continuous words}
}
\value{
A wordcloud image
}
\description{
create_wordcloud
}
\examples{
create_wordcloud(lyrics = "outputs/lyrics/Lyrics_TaylorSwift.json")
}
