% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_helper_functions.R
\name{download_full_discography}
\alias{download_full_discography}
\title{download_full_discography}
\usage{
download_full_discography(name_artist)
}
\arguments{
\item{name_artist}{Name of the artist}
}
\value{
Downloads a json file
}
\description{
download_full_discography
}
\examples{
download_full_discography("Tool")
}
