% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_individual_songs.R
\name{download_individual_songs}
\alias{download_individual_songs}
\title{download_individual_songs
Download all songs from a processed spotify list}
\usage{
download_individual_songs(
  processed_spotify_list,
  min_s_wait = 5,
  max_s_wait = 10
)
}
\arguments{
\item{processed_spotify_list}{A spotify list from get_spotify_list()}

\item{min_s_wait}{Min seconds to wait between requests}

\item{max_s_wait}{Max seconds to wait between requests}
}
\value{
.json files downloaded using the Genius API
}
\description{
download_individual_songs
Download all songs from a processed spotify list
}
\examples{
download_individual_songs(processed_spotify_list)
}
