% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_process_spotify_list.R
\name{download_process_spotify_list}
\alias{download_process_spotify_list}
\title{download_process_spotify_list}
\usage{
download_process_spotify_list(
  spotify_list_URL,
  only_new = FALSE,
  continue_after_error = FALSE,
  end_message = FALSE
)
}
\arguments{
\item{spotify_list_URL}{URL of Spotify list. For example: https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9}

\item{only_new}{Download only authors we don't have?}

\item{continue_after_error}{Will dowload authors not in the main folder}

\item{end_message}{Send message using {ntfy}}
}
\value{
A DF with the full processed data
}
\description{
download_process_spotify_list
}
\examples{
download_process_spotify_list("https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9")
}
