% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_helper_functions.R
\name{download_song}
\alias{download_song}
\title{download_song}
\usage{
download_song(name_artist, name_song, min_s_wait = 5, max_s_wait = 10)
}
\arguments{
\item{name_artist}{Name of the artist}

\item{name_song}{Name of the song}
}
\value{
Downloads a json file
}
\description{
download_song
}
\examples{
download_individual_songs("Tool", "46&2")
}
