% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_helper_functions.R
\name{get_and_process}
\alias{get_and_process}
\title{get_and_process
Get and process Spotify list html (create zip)}
\usage{
get_and_process(WEB)
}
\arguments{
\item{WEB}{Spotify list URL}
}
\value{
A DF
}
\description{
get_and_process
Get and process Spotify list html (create zip)
}
\examples{
get_and_process("https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9")
}
