% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_helper_functions.R
\name{get_and_process_spotify_list}
\alias{get_and_process_spotify_list}
\title{get_and_process_spotify_list
Get and process Spotify list html (create zip)}
\usage{
get_and_process_spotify_list(spotify_list_URL)
}
\arguments{
\item{spotify_list_URL}{Spotify list URL}
}
\value{
A DF
}
\description{
get_and_process_spotify_list
Get and process Spotify list html (create zip)
}
\examples{
get_and_process_spotify_list("https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9")
}
