% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_helper_functions.R
\name{get_individual_songs}
\alias{get_individual_songs}
\title{get_individual_songs}
\usage{
get_individual_songs(name_artist, name_song)
}
\arguments{
\item{name_artist}{Name of the artist}

\item{name_song}{Name of the song}
}
\value{
Downloads a json file
}
\description{
get_individual_songs
}
\examples{
get_individual_songs("Tool", "46&2")
}
