% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_helper_functions.R
\name{get_songs}
\alias{get_songs}
\title{get_songs}
\usage{
get_songs(name_artist)
}
\arguments{
\item{name_artist}{}
}
\value{
Downloads a json file
}
\description{
get_songs
}
\examples{
get_songs("Tool")
}
