% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_all_lyrics.R
\name{process_all_lyrics}
\alias{process_all_lyrics}
\title{process_all_lyrics}
\usage{
process_all_lyrics(
  daemons_artists = 10,
  daemons_songs = 5,
  write_output = FALSE,
  filename_output = NULL,
  language = NULL
)
}
\arguments{
\item{daemons_artists}{cores to process artists json lyrics}

\item{daemons_songs}{cores to process songs json lyrics}

\item{write_output}{TRUE/FALSE}

\item{language}{Filter by language? Most common: en, es, fr, pt, ko}
}
\value{
A DF with all artits and individual songs
}
\description{
process_all_lyrics
}
\examples{
process_all_lyrics(write_output = TRUE)
}
