% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_html.R
\name{process_html}
\alias{process_html}
\title{process_html}
\usage{
process_html(WEB, DEBUG = FALSE)
}
\arguments{
\item{WEB}{URL for Spotify list}

\item{DEBUG}{TRUE/FALSE, show songs in Console}
}
\value{
A list with a DF and a vector of ARTISTS
}
\description{
process_html
}
\examples{
process_html("https://open.spotify.com/playlist/37i9dQZEVXbNFJfN1Vw8d9")
}
