% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_new_lyrics.R
\name{process_new_lyrics}
\alias{process_new_lyrics}
\title{process_new_lyrics
Read lyrics in  "outputs/lyrics_to_process/" and combine them with full lyrics}
\usage{
process_new_lyrics(
  folder_lyrics = "outputs/lyrics_to_process/",
  daemons = 10,
  write_output = FALSE,
  filename_output = NULL,
  language = NULL
)
}
\arguments{
\item{folder_lyrics}{Location of json files}

\item{daemons}{cores to process json lyrics}

\item{write_output}{TRUE/FALSE}

\item{filename_output}{Name of output file}

\item{language}{Filter by language? Most common: en, es, fr, pt, ko}
}
\value{
A DF with all artits and individual songs
}
\description{
process_new_lyrics
Read lyrics in  "outputs/lyrics_to_process/" and combine them with full lyrics
}
\examples{
process_new_lyrics(write_output = TRUE)
}
