% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search words.R
\name{search_words}
\alias{search_words}
\title{search_words}
\usage{
search_words(data, highlight_words, n_sentences_around = 2)
}
\arguments{
\item{data}{File with lyrics songs (json or csv)}

\item{highlight_words}{Words to highlight}

\item{n_sentences_around}{How many sentences to show around the highlighted words}
}
\value{
A DT table
}
\description{
search_words
}
\examples{
 search_words(data = "outputs/lyrics/Lyrics_MykeTowers.json", highlight_words = "amor")
}
