% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate_paragraphs.R
\name{separate_paragraphs}
\alias{separate_paragraphs}
\title{separate_paragraphs
Separates a DF created with process_all_lyrics() into individual paragraphs
and cleans the paragraphs}
\usage{
separate_paragraphs(
  DF_lyrics,
  daemons = 5,
  write_output = FALSE,
  filename_output = NULL,
  language = NULL
)
}
\arguments{
\item{DF_lyrics}{DF_lyrics dataframe}

\item{daemons}{a DF created with process_all_lyrics()}

\item{write_output}{TRUE/FALSE}

\item{filename_output}{Name of the output file}

\item{language}{es, en, ...}
}
\value{
A DF with paragraphs
}
\description{
separate_paragraphs
Separates a DF created with process_all_lyrics() into individual paragraphs
and cleans the paragraphs
}
\examples{
DF_paragraphs = separate_paragraphs(DF_ALL_es)
}
