% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_words.R
\name{search_words}
\alias{search_words}
\title{search_words}
\usage{
search_words(lyrics, highlight_words, n_sentences_around = 2)
}
\arguments{
\item{lyrics}{JSON file with lyrics of DF_lyrics}

\item{highlight_words}{Words to highlight}

\item{n_sentences_around}{How many sentences to show around the highlighted words}
}
\value{
A DT table
}
\description{
search_words
}
\examples{
 search_words(lyrics = "outputs/lyrics_to_process/Lyrics_Tool.json", highlight_words = "amor")
}
